+++ src/lib.rs
@@ -29,6 +29,7 @@ mod mission;
 mod mission_location;
 mod on_change_strategy;
 mod report;
+mod scope;
 mod scroll;
 mod settings;
 mod state;
@@ -69,6 +70,7 @@ pub use {
     mission_location::*,
     on_change_strategy::*,
     report::*,
+    scope::*,
     scroll::*,
     settings::*,
     state::*,